# apps/control-plane/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenCrane Control Plane API",
    "version": "1.0.0",
    "description": "Multi-tenant AI agent platform management API. All resources require bearer-token authentication unless noted."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API prefix"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error description."
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "detail": {
            "type": "string",
            "description": "Optional extra context."
          }
        }
      },
...
      "Policy": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "tenantSelector": {
            "type": "object"
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "egressRules": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "mcpServers": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static bearer token. Pass as Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/policies": {
      "get": {
        "operationId": "listPolicies",
        "summary": "List all access policies",
        "tags": [
          "Policies"
        ],
        "responses": {
          "200": {
            "description": "Policy list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Policy"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createPolicy",
        "summary": "Create an access policy (dual-write: K8s CRD + database)",
        "tags": [
          "Policies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Policy created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/policies/repair": {
      "post": {
        "operationId": "repairPolicyProjection",
        "summary": "Repair AccessPolicy projection rows from CRD source of truth",
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "dryRun",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repair report.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/policies/{name}": {
      "get": {
        "operationId": "getPolicy",
        "summary": "Get a single access policy by name",
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Policy detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          },
          "404": {
            "description": "Policy not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updatePolicy",
        "summary": "Update an access policy",
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policy updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deletePolicy",
        "summary": "Delete an access policy",
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Policy deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}